requirement.txt
langchain
langchain-core

langchain-google-genai
google-generativeai


langchain-huggingface
transformers
huggingface-hub
sentence-transformers

python-dotenv

numpy
scikit-learn

flask